Move the clean up of legacy messages into a separate module
